+++ lib/src/cache/expiration_cache.dart
@@ -4,14 +4,15 @@ import 'abstract_cache.dart';
 
 /// A cache which evicts entries after a certain amount of time
 class ExpirationCache<K, V> extends AbstractCache<K, V> {
-  final Duration expiration;
+  final int expiration;
 
   int lastCleanup;
 
-  ExpirationCache(Storage<K, V>? storage, this.expiration, int capacity)
+  ExpirationCache({Storage<K, V>? storage, required Duration expiration, required int capacity})
       : assert(!expiration.isNegative),
         assert(expiration.inMilliseconds > 0),
         lastCleanup = DateTime.now().millisecondsSinceEpoch,
+        this.expiration = expiration.inMilliseconds,
         super(storage: storage, capacity: capacity);
 
   @override
@@ -20,11 +21,10 @@ class ExpirationCache<K, V> extends AbstractCache<K, V> {
   }
 
   @override
-  CacheEntry<K, V>? beforeGet(CacheEntry<K, V> entry) {
-    if ((entry as ExpirationCacheEntry).insertTime <
-        DateTime.now().millisecondsSinceEpoch - expiration.inMilliseconds) {
+  CacheEntry<K, V>? beforeGet(K key, CacheEntry<K, V> entry) {
+    if ((entry as ExpirationCacheEntry).insertTime < DateTime.now().millisecondsSinceEpoch - expiration) {
       // do not call onCapacity because if the entry is expired we do not want to keep it anyway
-      storage.removeInternal(entry.key);
+      storage.removeInternal(key);
       return null;
     }
     return entry;
@@ -33,12 +33,12 @@ class ExpirationCache<K, V> extends AbstractCache<K, V> {
   @override
   void onCapacity(K key, V element) {
     if (length < capacity) return;
-    int toRemove =
-        DateTime.now().millisecondsSinceEpoch - expiration.inMilliseconds;
+    int toRemove = DateTime.now().millisecondsSinceEpoch - expiration;
     if (lastCleanup > toRemove) return;
-    Iterable<CacheEntry<K, V>> itemsToRemove = storage.entries.where(
-        (element) => (element as ExpirationCacheEntry).insertTime < toRemove);
-    itemsToRemove.forEach((element) {
+    Iterable<MapEntry<K, CacheEntry<K, V>>> itemsToRemove =
+        storage.entries.entries.where((element) => (element.value as ExpirationCacheEntry).insertTime < toRemove);
+    // convert to list to avoid concurrent modification exceptions
+    itemsToRemove.toList().forEach((element) {
       // do not call onCapacity because if the entry is expired we do not want to keep it anyway
       storage.removeInternal(element.key);
     });
@@ -53,5 +53,5 @@ class ExpirationCacheEntry<K, V> extends CacheEntry<K, V> {
 
   ExpirationCacheEntry(K key, V? value)
       : insertTime = DateTime.now().millisecondsSinceEpoch,
-        super(key, value);
+        super(value);
 }