+++ lib/src/cache/lfu_cache.dart
@@ -5,17 +5,14 @@ import '../storage.dart';
 
 /// Least frequently used cache. Items which are not used often gets evicted first
 class LfuCache<K, V> extends AbstractCache<K, V> {
-  LfuCache({required Storage<K, V>? storage, required int capacity})
-      : super(storage: storage, capacity: capacity);
+  LfuCache({required Storage<K, V>? storage, required int capacity}) : super(storage: storage, capacity: capacity);
 
   @override
   void onCapacity(K key, V element) {
     if (length < capacity) return;
     // Iterate on all keys, so the eviction is O(n) to allow an insertion at O(1)
-    LfuCacheEntry<K, V> min = storage.entries
-        .map((e) => e as LfuCacheEntry<K, V>)
-        .reduce((element1, element2) =>
-            element1.use < element2.use ? element1 : element2);
+    MapEntry<K, CacheEntry<K, V>> min = storage.entries.entries
+        .reduce((element1, element2) => (element1.value as LfuCacheEntry).use < (element2.value as LfuCacheEntry).use ? element1 : element2);
 
     storage.onCapacity(min.key);
   }
@@ -26,7 +23,7 @@ class LfuCache<K, V> extends AbstractCache<K, V> {
   }
 
   @override
-  CacheEntry<K, V>? beforeGet(CacheEntry<K, V> entry) {
+  CacheEntry<K, V>? beforeGet(K key, CacheEntry<K, V> entry) {
     (entry as LfuCacheEntry).use++;
     return entry;
   }
@@ -37,5 +34,5 @@ class LfuCache<K, V> extends AbstractCache<K, V> {
 class LfuCacheEntry<K, V> extends CacheEntry<K, V> {
   int use = 0;
 
-  LfuCacheEntry(K key, V? value) : super(key, value);
+  LfuCacheEntry(K key, V? value) : super(value);
 }