+++ lib/src/cache/simple_cache.dart
@@ -6,12 +6,11 @@ import 'abstract_cache.dart';
 /// than appending keys in the storage, and remove first inserted keys when
 /// storage is full
 class SimpleCache<K, V> extends AbstractCache<K, V> {
-  SimpleCache({Storage<K, V>? storage, required int capacity})
-      : super(storage: storage, capacity: capacity);
+  SimpleCache({Storage<K, V>? storage, required int capacity}) : super(storage: storage, capacity: capacity);
 
   @override
   CacheEntry<K, V> createCacheEntry(K key, V value) {
-    return CacheEntry(key, value);
+    return CacheEntry(value);
   }
 
   @override