+++ lib/src/cache_entry.dart
@@ -1,6 +1,6 @@
+/// A cache entry. Note that the key type is defined though not used in the default entry. It could be used for more complex entries
 class CacheEntry<K, V> {
-  final K key;
   final V? value;
 
-  const CacheEntry(this.key, this.value);
+  const CacheEntry(this.value);
 }