+++ lib/src/storage/simple_storage.dart
@@ -5,8 +5,7 @@ import '../storage.dart';
 
 /// A simple storage class which is backed by a LinkedHashMap internally
 class SimpleStorage<K, V> implements Storage<K, V> {
-  final Map<K, CacheEntry<K, V>> _internalMap =
-      LinkedHashMap<K, CacheEntry<K, V>>();
+  final Map<K, CacheEntry<K, V>> _internalMap = LinkedHashMap<K, CacheEntry<K, V>>();
 
   /// if onEvict is set that method is called whenever an entry is removed from the cache.
   /// At the time the method is called the entry is already removed.
@@ -15,15 +14,10 @@ class SimpleStorage<K, V> implements Storage<K, V> {
   SimpleStorage({this.onEvict});
 
   @override
-  CacheEntry<K, V>? operator [](K key) {
-    return get(key);
-  }
-
-  @override
   void clear() {
     if (onEvict != null) {
-      _internalMap.values.forEach((element) {
-        if (element.value != null) onEvictInternal(element.key, element.value!);
+      _internalMap.entries.forEach((action) {
+        if (action.value.value != null) onEvictInternal(action.key, action.value.value!);
       });
     }
     _internalMap.clear();
@@ -40,21 +34,16 @@ class SimpleStorage<K, V> implements Storage<K, V> {
   }
 
   @override
-  void operator []=(K key, CacheEntry<K, V> value) {
-    set(key, value);
-  }
-
-  @override
   Storage set(K key, CacheEntry<K, V> value) {
     CacheEntry<K, V>? oldEntry = _internalMap[key];
     if (oldEntry != null && oldEntry.value != null && onEvict != null) {
-      onEvictInternal(oldEntry.key, oldEntry.value!);
+      onEvictInternal(key, oldEntry.value!);
     }
     setInternal(key, value);
     return this;
   }
 
-  void setInternal(K key, CacheEntry<K, V>  value) {
+  void setInternal(K key, CacheEntry<K, V> value) {
     _internalMap[key] = value;
   }
 
@@ -62,7 +51,7 @@ class SimpleStorage<K, V> implements Storage<K, V> {
   CacheEntry<K, V>? remove(K key) {
     CacheEntry<K, V>? oldEntry = _internalMap.remove(key);
     if (oldEntry != null && oldEntry.value != null && onEvict != null) {
-      onEvictInternal(oldEntry.key, oldEntry.value!);
+      onEvictInternal(key, oldEntry.value!);
     }
     return oldEntry;
   }
@@ -71,14 +60,18 @@ class SimpleStorage<K, V> implements Storage<K, V> {
   CacheEntry<K, V>? removeInternal(K key) {
     CacheEntry<K, V>? oldEntry = _internalMap.remove(key);
     if (oldEntry != null && oldEntry.value != null && onEvict != null) {
-      onEvictInternal(oldEntry.key, oldEntry.value!);
+      onEvictInternal(key, oldEntry.value!);
     }
     return oldEntry;
   }
 
   @override
   CacheEntry<K, V>? onCapacity(K key) {
-    return remove(key);
+    CacheEntry<K, V>? oldEntry = _internalMap.remove(key);
+    if (oldEntry != null && oldEntry.value != null && onEvict != null) {
+      onEvictInternal(key, oldEntry.value!);
+    }
+    return oldEntry;
   }
 
   @override
@@ -93,5 +86,5 @@ class SimpleStorage<K, V> implements Storage<K, V> {
   List<K> get keys => _internalMap.keys.toList();
 
   @override
-  List<CacheEntry<K, V>> get entries => _internalMap.values.toList();
+  Map<K, CacheEntry<K, V>> get entries => _internalMap;
 }